bug fixes in folder handling

+++ example.py
@@ -3,13 +3,13 @@ import numpy as np
 import time
 import os
 from scipy.stats import norm
-try:
+if os.path.exists('anomaly_scores'):
     if os.listdir('anomaly_scores') > 0:
         #remove all files in captures folder
         files = ['anomaly_scores/'+f for f in os.listdir('anomaly_scores')]
         for f in files:
             os.remove(f)
-except:
+else:
     os.mkdir('anomaly_scores')
 for temp_path in os.listdir('captures'):
     path = 'captures/'+temp_path
